requirements.txt
adbc-driver-manager==1.10.0
adbc-driver-postgresql==1.10.0
addict==1.0.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
click==8.3.1
connectorx==0.4.5
docutils==0.22.4
duckdb==1.4.4
faker==40.1.2
greenlet==3.3.1
h11==0.16.0
idna==3.11
importlib-resources==6.5.2
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
loro==1.10.3
marimo==0.19.7
markdown==3.10.1
markupsafe==3.0.3
msgspec==0.20.0
narwhals==2.15.0
numpy==2.4.2
packaging==26.0
pandas==3.0.0
parso==0.8.5
plotly==6.5.2
polars==1.37.1
polars-runtime-32==1.37.1
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pymdown-extensions==10.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlalchemy==2.0.46
sqlglot==28.10.0
sqlmodel==0.0.32
starlette==0.52.1
tomlkit==0.14.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
websockets==16.0
